Refactors and support for loading indicators, abillity to close arbitary tab
