+++ web/src/component_tabs.rs
@@ -1,21 +1,25 @@
 use crate::component_tab::Tab;
 use yew::prelude::*;
+use yew::virtual_dom::VChild;
 use yew::{ChildrenWithProps, Component, ComponentLink, Html, Properties};
 
 pub struct Tabs {
     link: ComponentLink<Self>,
-    selected: usize,
     props: Props,
 }
 
 #[derive(Properties, Clone, PartialEq)]
 pub struct Props {
     pub ontabselect: Callback<usize>,
+    pub ontabremove: Callback<usize>,
     pub children: ChildrenWithProps<Tab>,
+    pub selected: Option<usize>,
 }
 
+#[derive(Debug)]
 pub enum Msg {
     Selected(usize),
+    Removed(usize),
 }
 
 impl Component for Tabs {
@@ -23,18 +27,18 @@ impl Component for Tabs {
     type Properties = Props;
 
     fn create(props: Self::Properties, link: ComponentLink<Self>) -> Self {
-        Self {
-            link,
-            selected: 0,
-            props,
-        }
+        Self { link, props }
     }
 
     fn update(&mut self, msg: Self::Message) -> bool {
+        log::info!("TAB msg={:?}", msg);
         match msg {
             Msg::Selected(pos) => {
-                self.selected = pos;
                 self.props.ontabselect.emit(pos);
+                true
+            }
+            Msg::Removed(pos) => {
+                self.props.ontabremove.emit(pos);
                 true
             }
         }
@@ -42,15 +46,6 @@ impl Component for Tabs {
 
     fn change(&mut self, props: Self::Properties) -> bool {
         if self.props != props {
-            // If we have just added the first tab
-            if self.props.children.is_empty() && !props.children.is_empty() {
-                self.update(Msg::Selected(0));
-            }
-            // If we have just removed the current tab
-            //TODO: this wont work if we can remove any tab, selection will need to be tracked by parent
-            if self.props.children.len() > props.children.len() {
-                self.update(Msg::Selected(self.selected.checked_sub(1).unwrap_or(0)));
-            }
             self.props = props;
             true
         } else {
@@ -64,19 +59,43 @@ impl Component for Tabs {
                  <ul class="nav nav-tabs">
                       { for self.props.children.iter().enumerate().map(|(i, e)| html! {
                          <li class="nav-item" role="tablist">
-                             <a class={format!("nav-link {}", if self.selected == i {"active"} else {""})} role="tab" onclick=self.link.callback(move |_| Msg::Selected(i))>{&e.props.label}</a>
+                            <span class={format!("nav-link {}", if self.props.selected == Some(i) {"active"} else {""})} role="tab" onclick=self.link.callback(move |_| Msg::Selected(i))>
+                                <a style="vertical-align: middle;">{&e.props.label}</a>{ self.tab_details(e, i) }
+                             </span>
                          </li>
                       })}
                  </ul>
 
                  <div class="tab-content">
                       { for self.props.children.iter().enumerate().map(|(i, e)| html! {
-                         <div class={format!("tab-pane fade {}", if self.selected == i {"show active"} else {""})} role="tabpanel">
+                         <div class={format!("tab-pane fade {}", if self.props.selected == Some(i) {"show active"} else {""})} role="tabpanel">
                              {e}
                          </div>
                       })}
                  </div>
              </>
+        }
+    }
+}
+
+impl Tabs {
+    fn tab_details(&self, tab: VChild<Tab>, index: usize) -> Html {
+        if tab.props.loading {
+            self.loading_spinner()
+        } else {
+            self.remove_button(index)
+        }
+    }
+
+    fn remove_button(&self, index: usize) -> Html {
+        html! {
+            <span onclick=self.link.callback(move |_| Msg::Removed(index))><img src={"icon/x.svg"} style={"width: 24px; height: 24px;"} class={"mr-2"}/></span>
+        }
+    }
+
+    fn loading_spinner(&self) -> Html {
+        html! {
+            <span>{"Loading"}</span>
         }
     }
 }
